kbc: Handle setting typematic rate/delay
Per [1], use a default delay of 500ms and a rate of 10.9 CPS. When the
Linux kernel starts, this will be updated by atkbd to the fastest
settings possible [2]: 250ms delay and 30 CPS.

Modifying these values with kbdrate to test settings will cause two
spurious ACKs in the kernel. This behavior is also present in the
proprietary firmware.

[1]: https://web.archive.org/web/20180217074705/http://computer-engineering.org/ps2keyboard/
[2]: https://github.com/torvalds/linux/blob/c5f86891185c408b2241ba9a82ae8622d8386aff/drivers/input/keyboard/atkbd.c#L863-L868