galp4: Add numpad keys to Fn layer
Fn must be held to use the numpad, regardless of the state of NumLock.
This differs from the proprietary EC, which swaps the numpad and letters
when NumLock is enabled.

The OS determines if the numpad is used as navigation keys or number
keys, depending on the state of NumLock.