lemp11 fan curve
This is a proposal to add a fan curve for the Lemur Pro `lemp11`. I have been testing it for about a week and have seen significant improvements, approximately 20 degrees cooler, in temperatures both idling and under load. If this is too aggressive the values can always be changed, but I think it would be of value to add a curve that improves temps without adding much if any extra noise. So far I have not noticed a difference in terms of sound.