kbled: Add a "step" function for BKL hotkey
Replace the get+set logic with a step function to change the backlight
level for `K_KBD_BKL`.

Keyboards using a DAC have a different set of levels due to the
brightness difference between the keyboards.

NOTE: This introduces 0 as a level, meaning that the decicated
"backlight down" key on the 15" keyboards with keypads can now turn off
the backlight at the lowest level.

Signed-off-by: Tim Crawford <tcrawford@system76.com>