board/system76/common/battery.c: fix battery threshold handling
Charge only if the current battery level is lower than the start
threshold. Default thresholds (0, 100) do not matter, so skip checking
them.

Fixes the case where charging will unconditionally start if the end
threshold is set to 100. Also prevents unnecessary charging up to the
upper thresholds at power on if the current level is inbetween thresholds.

Signed-off-by: Michał Kopeć <michal.kopec@3mdeb.com>