Use distro provided rustup
This removes the need to download and run a shell script to set up the
environment. The distro is instead trusted to provide a valid binary.

Fedora has a package called "rustup", but it really is just the
rustup-init binary [1].

Additionally handle the behavior changes for rustup 1.28.0, which
removes implicit toolchain installation but adds support for installing
from `rust-toolchain.toml`.

[1]: https://src.fedoraproject.org/rpms/rustup

Signed-off-by: Tim Crawford <tcrawford@system76.com>