fan: Remove HEATUP, reduce COOLDOWN
Since these values actually use seconds for the unit again, their use
needs to be re-evaluated.

`HEATUP` can be removed entirely. We want the quick response to rising
temps without it instantly going to max from temp spikes. This is
handled by the new hardware control, which limits how quickly the fan
duty can change (3.9%/s).

`COOLDOWN` is still useful for hysteresis, to provide extra cooling as
temps come down, but does not need to run for 10/20s it is set to.

Signed-off-by: Tim Crawford <tcrawford@system76.com>