Add manual fan control
Allow fan target duties to be set via ACPI or SMFI command. This will
allow system firmware or OS to set fan duty, which can be used for
testing or implementing user-defined fan tables.

Setting PWM via SMFI already existed, but would not work as the value
would simply be overwritten by the EC.

RPM target is not supported.

Signed-off-by: Tim Crawford <tcrawford@system76.com>