Add Always on USB
Add option to partially power platform while off (VccPRIM and VDD5), in
order to keep the USB-A and USB-C ports powered. This allows using the
laptop as a power bank while it is off.

Refactor the power sequencing code to make it easier to partially power
the platform instead of only going full on or full off.

Power draw measured at USB-C plug is 0.161W, so we still get weeks of
standby.

Tested:
- V540TU: Both USB-C ports work, and the right USB-A port works in BC
  1.2 DCP mode. Left USB-A port does not work, but that's expected.
- NV41MZ: Only right USB-A port stays on. Tested unit does not have the
  USB charger option, so only slow charging works.

Signed-off-by: Michał Kopeć <michal.kopec@3mdeb.com>