usbpd: add PDC mode watchdog
As noted in issue https://github.com/Dasharo/dasharo-issues/issues/1505,
the PDC may sometimes get into a weird mode from which it can only
recover by issuing a cold reset. Add a watchdog that monitors the PDC
operational mode and kicks it if it has been in an unknown mode for more
than 1 second.

Signed-off-by: Michał Kopeć <michal.kopec@3mdeb.com>